seperated and modularized concerns
