+++ simple_server/server.js
@@ -6,3 +6,5 @@ app.listen(process.env.PORT || 3000, function(){
   console.log("Express server listening on port %d in %s mode", this.address().port, app.settings.env);
   console.log(__dirname);
 });
+
+module.exports = app;
